feat: dua_subtrees.sh script
@lorenzo-bianchi @giorgiomanca97
This wraps subtree interactions. It is meant to be used instead of git commands to interact with subtrees.
It is meant to be executed, with appropriate arguments, from both outside (during setup) and inside (during development) a container. There was no suitable way to define it so as to execute it from outside, and source it from inside.
It should work. Embeds many of the commands we tested but as with much of this stuff, only our future experience will tell.