fix: colcon-argcomplete sourcing
@alessandrotenaglia @giorgiomanca97 @ErLex20
FYI: colcon-argcomplete must be sourced and activated before ros2 is sourced, so that it can then activate ros2 shell completion.
This order is necessary only on SOME platforms.
To apply the patch in subsequent PRs: you may not recreate all the targets, just modify the ros2.sh file in the ones you already created.
The fix consists in moving the lines involved here up as shown here.